+++ src/Symfony/Component/DependencyInjection/Tests/Compiler/RegisterAutoconfigureAttributesPassTest.php
@@ -19,6 +19,8 @@ use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Reference;
 use Symfony\Component\DependencyInjection\Tests\Fixtures\AutoconfigureAttributed;
 use Symfony\Component\DependencyInjection\Tests\Fixtures\AutoconfiguredInterface;
+use Symfony\Component\DependencyInjection\Tests\Fixtures\FactoryAttributeService;
+use Symfony\Component\DependencyInjection\Tests\Fixtures\FactoryAttributeServiceAsString;
 use Symfony\Component\DependencyInjection\Tests\Fixtures\ParentNotExists;
 
 class RegisterAutoconfigureAttributesPassTest extends TestCase
@@ -47,6 +49,7 @@ class RegisterAutoconfigureAttributesPassTest extends TestCase
             ->addTag('another_tag', ['attr' => 234])
             ->addMethodCall('setBar', [2, 3])
             ->setBindings(['$bar' => $argument])
+            ->setFactory([null, 'create'])
         ;
         $this->assertEquals([AutoconfigureAttributed::class => $expected], $container->getAutoconfiguredInstanceof());
     }
@@ -87,5 +90,39 @@ class RegisterAutoconfigureAttributesPassTest extends TestCase
         (new RegisterAutoconfigureAttributesPass())->process($container);
 
         $this->addToAssertionCount(1);
+    }
+
+    public function testFactoryAttribute()
+    {
+        $container = new ContainerBuilder();
+        $container->register('foo', FactoryAttributeService::class)
+            ->setAutoconfigured(true);
+
+        $argument = new BoundArgument('foo', false, BoundArgument::INSTANCEOF_BINDING, realpath(__DIR__.'/../Fixtures/FactoryAttributeService.php'));
+
+        (new RegisterAutoconfigureAttributesPass())->process($container);
+
+        $expected = (new ChildDefinition(''))
+            ->setFactory([null, 'create'])
+            ->setBindings(['$foo' => $argument])
+        ;
+        $this->assertEquals([FactoryAttributeService::class => $expected], $container->getAutoconfiguredInstanceof());
+    }
+
+    public function testFactoryAttributeAsString()
+    {
+        $container = new ContainerBuilder();
+        $container->register('foo', FactoryAttributeServiceAsString::class)
+            ->setAutoconfigured(true);
+
+        $argument = new BoundArgument('foo', false, BoundArgument::INSTANCEOF_BINDING, realpath(__DIR__.'/../Fixtures/FactoryAttributeServiceAsString.php'));
+
+        (new RegisterAutoconfigureAttributesPass())->process($container);
+
+        $expected = (new ChildDefinition(''))
+            ->setFactory([null, 'create'])
+            ->setBindings(['$foo' => $argument])
+        ;
+        $this->assertEquals([FactoryAttributeServiceAsString::class => $expected], $container->getAutoconfiguredInstanceof());
     }
 }