+++ src/Symfony/Component/DependencyInjection/CHANGELOG.md
@@ -18,6 +18,7 @@ CHANGELOG
  * Add support for autowiring services as closures using `#[AutowireCallable]` or `#[AutowireServiceClosure]`
  * Deprecate `#[MapDecorated]`, use `#[AutowireDecorated]` instead
  * Deprecate the `@required` annotation, use the `Symfony\Contracts\Service\Attribute\Required` attribute instead
+ * Add `#[Factory]` and `factory` option to `#[Autoconfigure]`
 
 6.2
 ---